Use an equal check on the referent rather than on the string
This can come in handy in a case where the user has implemented a custom equals implementation that may return TRUE even if the resulting string is not the same, and it seems more logical to rather check the actual referent instead of the string representation.
Note that this check will never occur if not this and the to-be-compared-Result are not both on the same string referent state. (&& ends if left statement is false already)